+++ src/data/definitions/Settings.ts
@@ -1,0 +1,13 @@
+import { z } from "zod";
+
+export const SettingsSchema = z.object({
+  name: z.string(),
+  homePath: z.string(),
+});
+
+export type Settings = z.infer<typeof SettingsSchema>;
+
+export const defaultSettings: Settings = {
+  name: "Web App",
+  homePath: "home",
+};
