+++ src/data/definitions/index.ts
@@ -1,0 +1,6 @@
+export * from "./Block";
+export * from "./EditorData";
+export * from "./Page";
+export * from "./PageMetaData";
+export * from "./Settings";
+export * from "./WebAppConfig";
